% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHeatMap.R
\name{createHeatMap}
\alias{createHeatMap}
\title{Create Heat Map}
\usage{
createHeatMap(score_df, y.var, x.var)
}
\arguments{
\item{score_df}{data frame of scores of each signaling pathway.}

\item{y.var}{y variable specified by the user.}

\item{x.var}{x variable specified by the user.}
}
\value{
Heat Map of the scores.
}
\description{
Creates Heat Map using ggplot2 with the scores.
}
