% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPathway.R
\name{extractPathway}
\alias{extractPathway}
\title{Extract Signaling Pathways}
\usage{
extractPathway(gene)
}
\arguments{
\item{gene}{any gene.}
}
\value{
list of pathways.
}
\description{
Queries the PathCards website to find signaling pathways for each gene.
}
