% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixNames.R
\name{fixNames}
\alias{fixNames}
\title{Fix Pathway Names}
\usage{
fixNames(names_list)
}
\arguments{
\item{names_list}{list of pathway names.}
}
\value{
modified list of pthway names.
}
\description{
Delete any unnecessary symbols within pathway names.
}
