% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryPathcardsFunction.R
\name{queryPathcards}
\alias{queryPathcards}
\title{Query Pathcards Website}
\usage{
queryPathcards(gene_list)
}
\arguments{
\item{gene_list}{list of genes.}
}
\value{
dataframe with original genes and corresponding signaling pathways and a list of signaling pathways and corresponding gene lists.
}
\description{
Finds signaling pathways for each gene and then finds gene list for each signaling pathway.
}
