% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/receptorLigandAnnotationFromObject.R
\name{receptor.ligand.annotation.from.object}
\alias{receptor.ligand.annotation.from.object}
\title{Receptor Ligand Annotation From Seurat Object}
\usage{
receptor.ligand.annotation.from.object(seurat.object, group.by,
  receptor.ligand.mat, mean.expression.theshold = 0.1,
  percent.expression.threshold = 0.2)
}
\arguments{
\item{seurat.object}{seurat object set by the user.}

\item{group.by}{variables to group by set by the user.}

\item{receptor.ligand.mat}{the list of receptors and ligands.}

\item{percent.expression.threshold}{percent expression threshold set by the user.}

\item{mean.expression.threshold}{mean expression threshold set by the user.}
}
\value{
list of ligand-receptor pairs with expression levels.
}
\description{
Uses the seurat object to identify certain ligand-receptor pairs who have expression levels higher than the ones set by the user.
}
