% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/receptorLigandPairAnnotation.R
\name{receptor.ligand.pair.annotation}
\alias{receptor.ligand.pair.annotation}
\title{Annotate Receptor Ligand Pair}
\usage{
receptor.ligand.pair.annotation(Mat1, Mat2, L.R.Mat, group.by)
}
\description{
Subfunction for the function receptor.ligand.annotation.from.object
}
