% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreGenes.R
\name{scoreGenes}
\alias{scoreGenes}
\title{Score Genes}
\usage{
scoreGenes(result_list, seurat_object, group_by_names)
}
\arguments{
\item{result_list}{list created by pathcardsQuery function}

\item{seurat_object}{seurat object specified by user}

\item{group_by_names}{specific categories user specifies the scores to be grouped by}
}
\value{
list of pathways with their entire score and pathways with their summarized scores
}
\description{
Scores each pathway with a specificed seurat object
}
